requirements.txt
django
djangorestframework
psycopg2-binary
shapely
django_ratelimit